requirements.txt
# pip install -r requirements.txt  -i https://pypi.tuna.tsinghua.edu.cn/simple
pytest
coverage
fastapi[all]~=0.114.0
uvicorn~=0.18.2
logbook~=1.7.0.post0
pymysql
peewee~=3.17.6
pathlib
jieba
nltk
scrapy
pandas
threadpool
requests
snownlp
geopy~=2.4.1



pydantic~=1.10.12
PyJWT~=2.10.1
flask~=3.1.0
werkzeug~=3.1.3
numpy~=2.1.1
starlette~=0.38.5